requirements.txt
flask==3.1.1 
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.1.0
PyMySQL==1.1.0
cryptography
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
Flask-WTF==1.2.1
WTForms==3.1.1
email_validator==2.1.1
langchain==0.0.340
langchain-google-genai==0.0.6
google-generativeai==0.3.2
pandas>=2.2.0
numpy>=1.26.0
chromadb==0.4.18
langgraph>=1.0.3
# Production hardening (Week 5)
pydantic>=2.0.0
# Testing
pytest>=8.0.0
pytest-flask>=1.3.0
# Missing transitive deps (already used in codebase)
sentence-transformers>=2.2.0
flask-socketio>=5.3.0
scikit-learn>=1.3.0
joblib>=1.3.0
